+++ connection_tuple.h
@@ -11,16 +11,16 @@
 
 struct connection_tuple
 {
-    connection_tuple(uint64_t source_address, uint8_t source_port, uint64_t destination_address, uint8_t destination_port);
+    connection_tuple(uint64_t source_address, uint16_t source_port, uint64_t destination_address, uint16_t destination_port);
 
     std::string to_string() const;
 
     bool operator==(const connection_tuple &rhs) const;
 
     uint64_t source_address;
-    uint8_t source_port;
+    uint16_t source_port;
     uint64_t destination_address;
-    uint8_t destination_port;
+    uint16_t destination_port;
     // TODO: message_type?
 };
